Bump to version 0.0.706-DRAFT
- Add and use method putIfAbsent to RateCache
- Rename method record to increment
- Add argument amount to method Rate#increment
- Remove exception RateLimitExceededException from Rate#increment method
- Rename DefaultRateLimiter to SimpleRateLimiter
- Remove CompositeRate and related classes
- Re-design RateLimitHandler to improve flexibility

+++ src/main/java/com/looseboxes/ratelimiter/annotation/RateLimitProcessor.java
@@ -49,7 +49,7 @@ public class RateLimitProcessor extends AbstractProcessor {
         return false;
     }
 
-    public static String getErrorMessageIfInvalidLimit(int limit, String resultIfNone) {
+    public static String getErrorMessageIfInvalidLimit(long limit, String resultIfNone) {
         return limit < 0 ? "Invalid limit: " + limit : resultIfNone;
     }
